+++ specreduce/tests/test_background.py
@@ -2,7 +2,7 @@ import pytest
 import numpy as np
 
 import astropy.units as u
-from astropy.nddata import CCDData
+from astropy.nddata import VarianceUncertainty
 from specutils import Spectrum1D
 
 from specreduce.background import Background
@@ -16,7 +16,11 @@ from specreduce.tracing import FlatTrace, ArrayTrace
 image = np.ones(shape=(30, 10))
 for j in range(image.shape[0]):
     image[j, ::] *= j
-image = CCDData(image, unit=u.Jy)
+image = Spectrum1D(image * u.DN,
+                   uncertainty=VarianceUncertainty(np.ones_like(image)))
+image_um = Spectrum1D(image.flux,
+                      spectral_axis=np.arange(image.data.shape[1]) * u.um,
+                      uncertainty=VarianceUncertainty(np.ones_like(image.data)))
 
 
 def test_background():
@@ -28,12 +32,21 @@ def test_background():
     trace = FlatTrace(image, trace_pos)
     bkg_sep = 5
     bkg_width = 2
-    # all the following should be equivalent:
+
+    # all the following should be equivalent, whether image's spectral axis
+    # is in pixels or physical units:
     bg1 = Background(image, [trace-bkg_sep, trace+bkg_sep], width=bkg_width)
     bg2 = Background.two_sided(image, trace, bkg_sep, width=bkg_width)
     bg3 = Background.two_sided(image, trace_pos, bkg_sep, width=bkg_width)
-    assert np.allclose(bg1.bkg_array, bg2.bkg_array)
-    assert np.allclose(bg1.bkg_array, bg3.bkg_array)
+    assert np.allclose(bg1.bkg_image().flux, bg2.bkg_image().flux)
+    assert np.allclose(bg1.bkg_image().flux, bg3.bkg_image().flux)
+
+    bg4 = Background(image_um, [trace-bkg_sep, trace+bkg_sep], width=bkg_width)
+    bg5 = Background.two_sided(image_um, trace, bkg_sep, width=bkg_width)
+    bg6 = Background.two_sided(image_um, trace_pos, bkg_sep, width=bkg_width)
+    assert np.allclose(bg1.bkg_image().flux, bg4.bkg_image().flux)
+    assert np.allclose(bg1.bkg_image().flux, bg5.bkg_image().flux)
+    assert np.allclose(bg1.bkg_image().flux, bg6.bkg_image().flux)
 
     # test that creating a one_sided background works
     Background.one_sided(image, trace, bkg_sep, width=bkg_width)
@@ -45,8 +58,15 @@ def test_background():
     sub1 = image - bg1
     sub2 = bg1.sub_image(image)
     sub3 = bg1.sub_image()
-    assert np.allclose(sub1, sub2)
-    assert np.allclose(sub1, sub3)
+    assert np.allclose(sub1.flux, sub2.flux)
+    assert np.allclose(sub2.flux, sub3.flux)
+
+    sub4 = image_um - bg4
+    sub5 = bg4.sub_image(image_um)
+    sub6 = bg4.sub_image()
+    assert np.allclose(sub1.flux, sub4.flux)
+    assert np.allclose(sub4.flux, sub5.flux)
+    assert np.allclose(sub5.flux, sub6.flux)
 
     bkg_spec = bg1.bkg_spectrum()
     assert isinstance(bkg_spec, Spectrum1D)
@@ -54,7 +74,7 @@ def test_background():
     assert isinstance(sub_spec, Spectrum1D)
     # test that width==0 results in no background
     bg = Background.two_sided(image, trace, bkg_sep, width=0)
-    assert np.all(bg.bkg_image() == 0)
+    assert np.all(bg.bkg_image().flux == 0)
 
 
 def test_warnings_errors():